#15 PronunciationAcitivity 패키지 분리
