+++ app/src/main/java/com/example/androidlab2/RecordingActivity.kt
@@ -11,6 +11,7 @@ import androidx.appcompat.app.AlertDialog
 import androidx.appcompat.app.AppCompatActivity
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
+import com.example.androidlab2.ui.PronunciationActivity
 import java.io.File
 import java.text.SimpleDateFormat
 import java.util.*
