+++ app/src/main/java/com/example/androidlab2/model/PronunciationResult.kt
@@ -1,0 +1,5 @@
+package com.example.androidlab2.model
+
+data class PronunciationResult(
+    val result: String
+)
